fix rebase merge conflicts
