+++ package.json
@@ -12,12 +12,14 @@
   "dependencies": {
     "@google/local-home-sdk": "^1.0.0",
     "opc": "^1.1.3",
+    "ts-node": "^8.10.2",
     "yargs": "^15.3.1"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^3.3.0",
     "@typescript-eslint/parser": "^3.3.0",
     "ava": "^2.1.0",
+    "cbor": "^5.0.2",
     "eslint": "^7.2.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^6.11.0",
